Create encryption.py
The encryption program should accept commands given as lines via standard input. The first word (sequence of non-whitespace characters) should be treated as a command, and the rest of the line(after the first space) as the argument for that command. Output is printed to standard out as a line of text where the first word is a response type. The currently set key is remembered by the encryption program. The encryption program should handle the following commands.