Update driver.py
added the decrypt, history, and quit function:

decrypt: Provide the user with the option of using a string in the history or entering a new string. If a new string will be used, prompt the user for a string, record it in the history, and
 send an decrypt command to the encryption program. If the history will be used, provide the user with a menu where a number can be used to select a string stored in the history.
The results should be printed to standard output and saved in the history.

history: Show the history.

quit: Send QUIT to the encryption program and logger, and then exit the program.